refactor: account balance handling; change balance type to BigDecimal and update related methods for deposit and withdrawal

+++ src/main/java/com/Isabela01vSilva/bank_isabela/controller/response/account/AccountWithCustomerResponse.java
@@ -3,6 +3,7 @@ package com.Isabela01vSilva.bank_isabela.controller.response.account;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountStatus;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record AccountWithCustomerResponse(
@@ -12,7 +13,7 @@ public record AccountWithCustomerResponse(
         String numeroConta,
         AccountType tipoConta,
         AccountStatus statusConta,
-        Double saldo,
+        BigDecimal saldo,
         LocalDate dataCriacao
 ) {
 }
